requirements.txt
streamlit
anthropic
psycopg2-binary
sentence-transformers
python-dotenv